+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - Use bract.core 0.6.2-beta1
 - Add DEV logger functions `bract.ring.dev/log-*` for `bract.ring.wrapper/traffic-log-wrapper`
-  - [Todo] Add DEV mode logger config in `config.dev.edn` file
+  - Add DEV mode logger config in `config.dev.edn` file
 - Add support for [nginx-clojure-embedded](https://github.com/nginx-clojure/nginx-clojure/tree/master/nginx-clojure-embed)
   - `bract.ring.server/start-nginx-clojure-embedded-server`